+++ tests/test_unit/test_plugins/test_msgspec/test_validation.py
@@ -1,4 +1,5 @@
 import json
+from collections.abc import AsyncIterator
 from http import HTTPMethod, HTTPStatus
 from typing import final
 
@@ -14,6 +15,7 @@ from faker import Faker
 from inline_snapshot import snapshot
 
 from dmr import Body, Controller, ResponseSpec, modify, validate
+from dmr.exceptions import EndpointMetadataError
 from dmr.plugins.msgspec import MsgspecSerializer
 from dmr.test import DMRRequestFactory
 
@@ -206,3 +208,15 @@ def test_msgspec_struct_renames_work(
     assert response.status_code == HTTPStatus.CREATED, response.content
     assert response.headers == {'Content-Type': 'application/json'}
     assert json.loads(response.content) == request_data
+
+
+def test_msgspec_rejects_async_gen() -> None:
+    """Ensure msgspec controllers cannot define async generator endpoints."""
+    with pytest.raises(
+        EndpointMetadataError,
+        match='is an async generator',
+    ):
+
+        class _BadController(Controller[MsgspecSerializer]):
+            async def get(self) -> AsyncIterator[int]:
+                yield 1  # pragma: no cover
